Bug fixes
Fixed an error where position was declared incorrectly as a constant rather than let. Also, fixed an issue where the "require cycle" warning was showing up, which is usually displayed when there is a circular dependency between two or more modules in the code. In this case, it seemed that there was a circular dependency between LowerPanel.jsx and Facilities.jsx, which was causing the warning.